requirements.txt
customtkinter==5.2.0
packaging
groq
python-dotenv
